backend/requirements.txt
fastapi
uvicorn[standard]
asyncpg
sqlalchemy[asyncio]
psycopg2-binary
python-dotenv
